elab/webgpio/requirements.txt
cachelib==0.1.1
certifi==2018.8.24
click==7.1.2
csvalidate==1.1.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-CSV==1.2.0
Flask-Session==0.3.2
future==0.18.2
importlib-metadata==4.0.1
iso8601==0.1.14
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
numpy==1.19.5
pandas==0.23.4
parso==0.7.1
pfeiffer-vacuum-protocol==0.4
pluggy==0.13.1
pyserial==3.5
python-dateutil==2.6.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2018.5
PyYAML==5.4.1
RPi.GPIO==0.7.0
six==1.15.0
typing-extensions==3.10.0.0
ujson==4.0.2
Werkzeug==1.0.1
zipp==3.4.1
